requirements.txt
# PyTorch with MPS support for Apple Silicon
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Core ML and model libraries
transformers>=4.36.0
diffusers>=0.25.0
accelerate>=0.25.0
optimum>=1.16.0

# Memory optimization
xformers>=0.0.22
bitsandbytes>=0.42.0

# Video processing
opencv-python>=4.8.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
av>=10.0.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0

# Model hub and utilities
huggingface-hub>=0.19.0
safetensors>=0.4.0
omegaconf>=2.3.0
einops>=0.7.0
sentencepiece>=0.1.99

# Progress and visualization
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Memory profiling
psutil>=5.9.0
memory-profiler>=0.61.0

# Additional utilities for Apple Silicon
coremltools>=7.0
protobuf>=3.20.0